pyproject.toml
[project]
name = "asp"
version = "0.0.0"
description = "Attention Superlinéaire Polymorphe — projet de recherche"
readme = "DOC/README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [{ name = "synthetic-attention" }]

# --- Runtime ---
# Versions cibles : voir OPS/env/STACK.md.
# torch est intentionnellement absent des deps principales et listé dans
# les extras `cpu` ou `cuda` ci-dessous. Il faut activer l'un des deux
# pour avoir un environnement complet (cf. setup_env.sh).
dependencies = [
  "lightning>=2.4.0",
  "hydra-core>=1.3.2",
  "mlflow>=2.16.0",
  "numpy>=1.26.0,<3.0",
  "scipy>=1.13.0",
  "matplotlib>=3.9.0",
  "pandas>=2.2.0",
  "pyyaml>=6.0",
  "tqdm>=4.66.0",
]

[project.optional-dependencies]
# `uv sync --extra cpu` sur VPS sans GPU (torch CPU depuis PyPI)
# `uv sync --extra cuda` sur pod RunPod RTX 5090 (torch CUDA 12.8)
cpu = ["torch>=2.6.0"]
cuda = ["torch>=2.6.0"]

dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.6",
  "mypy>=1.11",
  "pre-commit>=3.8",
]

# Pas de [build-system] : le projet n'est pas un package Python distribuable
# pour l'instant. Le code vit dans CODE/ et est importé via PYTHONPATH (cf.
# OPS/scripts/setup_env.sh). uv gère uniquement les dépendances.

# ============================================================
# uv : configuration générale + indexes et sources pour wheels CUDA
# ============================================================

[tool.uv]
# Empêche uv de traiter ce projet comme un package à build/install.
# Les dépendances sont seules pertinentes ; le code vit dans CODE/.
package = false
# La logique : par défaut torch vient de PyPI (CPU). Si l'extra `cuda` est
# activé, torch vient de l'index PyTorch CUDA 12.8.
#
# IMPORTANT : si l'index `cu128` n'existe pas encore au moment d'installer
# (PyTorch publie progressivement les wheels Blackwell), basculer sur
# `cu126` ou `cu124` dans cette section et regénérer uv.lock.
# ============================================================

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128", extra = "cuda" },
]

# ============================================================
# Ruff (lint + format)
# ============================================================
[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "N", "SIM"]
ignore = []

[tool.ruff.format]
quote-style = "double"

# ============================================================
# Pytest
# ============================================================
[tool.pytest.ini_options]
testpaths = ["CODE"]
python_files = ["test_*.py"]
addopts = "-ra --strict-markers"
pythonpath = ["CODE"]

# ============================================================
# Mypy (strict, mais permissif sur libs sans stubs)
# ============================================================
[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
